Swapped API versions to be u64

+++ bspsuite-extinterface/src/apis/api_info.rs
@@ -3,12 +3,12 @@ use bspffi::types::XCStr;
 pub struct ApiInfo
 {
 	pub name: XCStr<'static>,
-	pub version: usize,
+	pub version: u64,
 }
 
 impl ApiInfo
 {
-	pub const fn new(name: &'static str, version: usize) -> Self
+	pub const fn new(name: &'static str, version: u64) -> Self
 	{
 		return Self {
 			name: XCStr::new(name),
